+++ wallet/wallet-controller/src/runtime_wallet.rs
@@ -102,6 +102,19 @@ where
         }
     }
 
+    pub fn reset_inmempool_txs_to_inactive(
+        &mut self,
+        wallet_events: Option<&impl WalletEvents>,
+    ) -> WalletResult<()> {
+        match self {
+            RuntimeWallet::Software(w) => w.reset_inmempool_txs_to_inactive(wallet_events),
+            #[cfg(feature = "trezor")]
+            RuntimeWallet::Trezor(w) => w.reset_inmempool_txs_to_inactive(wallet_events),
+            #[cfg(feature = "ledger")]
+            RuntimeWallet::Ledger(w) => w.reset_inmempool_txs_to_inactive(wallet_events),
+        }
+    }
+
     pub fn find_account_destination(&self, acc_outpoint: &AccountOutPoint) -> Option<Destination> {
         match self {
             RuntimeWallet::Software(w) => w.find_account_destination(acc_outpoint),